Plot customization

Parameters used to choose what to plot and how. These parameters are given to plot.SCGLR and pairs.SCGLR.

Details

Parameter name can be abbreviated (e.g. pred.col will be understood as predictors.color).
Options can be set globally using options("plot.SCGLR"). It will then provide default values that can be further overriden by giving explicit parameter value.

parameter name	type (default value). Description.

title	string (NULL). Main title of plot (override built-in).

labels.auto	logical (TRUE). Should covariate or predictor labels be aligned with arrows.

labels.offset	numeric (0.01). Offset by which labels should be moved from tip of arrows.

labels.size	numeric (1). Relative size for labels. Use it to globally alter label size.

expand	numeric (1). Expand factor for windows size. Use it for example to make room for clipped labels.

threshold	numeric. All covariates and/or predictors whose sum of square correlations with the two components of the plane lower than this threshold will be ignored.

observations	logical (FALSE). Should we draw observations.

observations.size	numeric (1). Point size.

observations.color	character ("black"). Point color.

observations.alpha	numeric (1). Point transparency.

observations.factor	logical (FALSE). Paint observations according to factor (specify factor).

predictors	logical or array of characters (FALSE). Should we draw predictors and optionally which one (TRUE means all).

predictors.color	string ("red"). Base color used to draw predictors.

predictors.alpha	numeric (1). Overall transparency for predictors (0 is transparent, 1 is opaque).

predictors.arrows	logical (TRUE). Should we draw arrows for predictors.

predictors.arrows.color	string (predictors.color). Specific color for predictor arrows.

predictors.arrows.alpha	numeric (predictors.alpha). Transparency for predictor arrows.

predictors.labels	logical (TRUE). Should we draw labels for predictors.

predictors.labels.color	string (predictors.color). Specific color for predictor labels.

predictors.labels.alpha	numeric (predictors.alpha). Transparency for predictor labels.

predictors.labels.size	numeric (labels.size). Specific size for predictor labels.

predictors.labels.auto	logical (labels.auto). Should predictor labels be aligned with arrows.

covariates	logical or array of characters (TRUE). Should we draw covariates and optionally which one (TRUE means all).

covariates.color	string ("black"). Base color used to draw covariates.

covariates.alpha	numeric (1). Overall transparency for covariates (0 is transparent, 1 is opaque).

covariates.arrows	logical (TRUE). Should we draw arrows for covariates.

covariates.arrows.color	string (covariates.color). Specific color for covariate arrows.

covariates.arrows.alpha	numeric (covariates.alpha). Transparency for covariate arrows.

covariates.labels	logical (TRUE). Should we draw labels for predictors.

covariates.labels.color	string (covariates.color). Specific color for predictor labels.

covariates.labels.alpha	numeric (covariates.alpha). Transparency for covariate labels.

covariates.labels.size	numeric (labels.size). Specific size for covariate labels.

covariates.labels.auto	logical (labels.auto). Should covariate labels be aligned with arrows.

factor	logical or character (FALSE). Should we draw a factor chosen among additionnal variables (TRUE mean first one).

factor.points	logical (TRUE). Should symbol be drawn for factors.

factor.points.size	numeric (4). Symbol size.

factor.points.shape	numeric (13). Point shape.

factor.labels	logical (TRUE). Should factor labels be drawn.

factor.labels.color	string ("black"). Color used to draw labels.

factor.labels.size	numeric (labels.size). Specific size for factor labels.

parameter name

Examples

# NOT RUN {
# setting parameters
plot(genus.scglr)
plot(genus.scglr, predictors=TRUE)
plot(genus.scglr, predictors=TRUE, pred.arrows=FALSE)

# setting global style
options(plot.SCGLR=list(predictors=TRUE, pred.arrows=FALSE))
plot(genus.scglr)

# setting custom style
myStyle <- list(predictors=TRUE, pred.arrows=FALSE)
plot(genus.scglr, style=myStyle)
# }