Regularization criterion types

Regularization criterion types

Usage

methodSR(
  phi = "vpi",
  l = 1,
  s = 1/2,
  maxiter = 1000,
  epsilon = 1e-06,
  bailout = 10
)

Arguments

phi

character string describing structural relevance used in the regularization process. Allowed values are "vpi" for Variable Powered Inertia and "cv" for Component Variance. Default to "vpi".

l

is an integer argument (>1) tuning the importance of variable bundle locality.

s

is a numeric argument (in [0,1]) tuning the strength of structural relevance with respect to goodness of fit.

maxiter

integer for maximum number of iterations of SR function

epsilon

positive convergence threshold

bailout

integer argument

Value

an object of class method.SCGLR