SCGLR generic plot

SCGLR generic plot

Usage

# S3 method for SCGLR
plot(x, ..., style = getOption("plot.SCGLR"), plane = c(1, 2))

Arguments

x

an object from SCGLR class.

...

optional arguments (see customize).

style

named list of values used to customize the plot (see customize)

plane

a size-2 vector (or string with separator) indicating which components are plotted (eg: c(1,2) or "1,2" or "1/2").

Value

an object of class ggplot.

Examples

if (FALSE) {
library(SCGLR)

# load sample data
data(genus)

# get variable names from dataset
n <- names(genus)
ny <- n[grep("^gen",n)]    # Y <- names that begins with "gen"
nx <- n[-grep("^gen",n)]   # X <- remaining names

# remove "geology" and "surface" from nx
# as surface is offset and we want to use geology as additional covariate
nx <-nx[!nx%in%c("geology","surface")]

# build multivariate formula
# we also add "lat*lon" as computed covariate
form <- multivariateFormula(ny,c(nx,"I(lat*lon)"),c("geology"))

# define family
fam <- rep("poisson",length(ny))

genus.scglr <- scglr(formula=form,data = genus,family=fam, K=4,
 offset=genus$surface)

summary(genus.scglr)

barplot(genus.scglr)

plot(genus.scglr)

plot(genus.scglr, predictors=TRUE, factor=TRUE)

pairs(genus.scglr)

}